requirements.txt
PyTurboJPEG @ git+https://github.com/lilohuang/PyTurboJPEG.git@c491d2ad562624fee834f51ef09c2b138e6636b3
modelhub-client @ git+https://github.com/ria-com/modelhub-client.git@d155a12e1c32a0a65b98c911f1e9e8786d02d93f
supervision==0.1.0
pycocotools>=2.0.2
onnx>=1.13.0
onnx-simplifier==0.4.10
fastai==2.7.10
GitPython==3.1.18
nltk==3.8.1
numpy==1.26.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.76
lmdb==1.4.1
lxml==4.9.4
basicsr==1.4.2
ultralytics==8.0.229
gevent==23.9.0.post1
pytorch-lightning==1.8.6
termcolor==2.4.0
paddleocr==2.6.1.3
paddlepaddle==2.5.1
timm==0.9.7
tqdm==4.65.0
uvicorn==0.26.0
fastapi==0.109.0
python-multipart==0.0.6
aiofiles
thop==0.1.1.post2209072238
loguru==0.7.2
SQLAlchemy==2.0.19
alembic==1.11.1
python-decouple==3.8
psycopg2-binary
asyncpg
pydantic-settings
facexlib
gfpgan
websockets